moving commands into CLI in response to suggestion from eduardo
