+++ CLI/singularity.go
